Refactor byte operations
